Move smoothing functions to libs/trend.ts and fix import bug
