+++ src/datasets/user-workday-holiday-linear.dataset.ts
@@ -7,10 +7,7 @@ import { createLineChartMeta } from './metas/line-chart.meta';
 import { PreferenceMeta } from '../services/preference/types';
 import { DAY } from '../utils/timeUnits';
 import { combineQuerySummariesFactories } from './summarizations/utils/commons';
-import * as WorkdayHolidayAbsoluteSummarization from './summarizations/workday-holiday-absolute.summarization';
-import * as WorkdayHolidayRelativeSummarization from './summarizations/workday-holiday-relative.summarization';
-import * as TrendSummarization from './summarizations/trend.summarization';
-import * as TrendRegressionSummarization from './summarizations/trend-regression.summarization';
+import * as TrendPartialSummarization from './summarizations/trend-partial.summarization';
 
 
 export interface Config {
@@ -84,8 +81,7 @@ export function create(config: Config): Dataset {
   const dataCube = generateCube(categories, measures, generateCubeConfig);
 
   const visitCountQuerySummariesFactory = combineQuerySummariesFactories(
-    TrendSummarization.queryFactory,
-    TrendRegressionSummarization.queryFactory,
+    TrendPartialSummarization.queryFactory,
   );
 
   const lineChartMeta = createLineChartMeta(
